# from csv: stream input, fix memory consumption