--- a/crates/nu-command/src/formats/from/tsv.rs
+++ b/crates/nu-command/src/formats/from/tsv.rs
@@ -139,6 +139,7 @@ fn from_tsv(
 
     let config = DelimitedReaderConfig {
         separator: '\t',
+        record_separator: '\n',
         comment,
         quote,
         escape,
@@ -148,7 +149,7 @@ fn from_tsv(
         trim,
     };
 
-    from_delimited_data(config, input, name)
+    from_delimited_data(config, input, name, engine_state.ctrlc.clone())
 }
 
 #[cfg(test)]