# Description

Fixes memory consumption of `from csv` by changing it to steam its input and output, rather than soak up everything into memory.

This is accomplished with a new type `StreamReader` that implements `std::io::Read` to give direct access to the bytes of the underlying values, if the values are a type that can give back bytes in a semantically unambiguous way. Types that cannot yield an error in the reads.

This reader is used as the input source of the `csv::Reader` so that it can be properly steamed.


### User-Facing Changes

The main benefit is the ability to stream an arbitrarily large input without exhausting memory.

In addition, the allowed input types have been expanded to allow literal `Value`s as well, e.g. these now work:

```nu
open --raw data.csv | lines | str replace 'foo' 'bar' | from csv
["foo,bar,baz", "quux,blerp,blorp"] | from csv
```

# Tests + Formatting

To test the memory consumption, I used [this publicly available dataset of registered electric vehicles](https://catalog.data.gov/dataset/electric-vehicle-population-data), using the first 1,000, 10,000, and 100,000 entries, in that order. I ran the following script to benchmark the changes against the current stable release:

```nu
#!/usr/bin/nu

let new_shellbin = $"($env.HOME)/src/nushell/target/aarch64-linux-android/release/nu"
let current_shellbin = "nu"
const dbpath = 'Electric_Vehicle_Population_Data.csv'

[$current_shellbin, $new_shellbin]
  | each {|shellbin|
    [1000, 10000, 100000]
      | each {|rows|
          do {
            time -f '%M %e %U %S' $shellbin -c (
              $"cat ($dbpath) | head -n ($rows) | from csv"
            )
          }
          | complete
          | get stderr
          | str trim
          | parse '{rss_max} {real} {user} {kernel}'
          | update cells -c [rss_max] { $"($in)kb" | into filesize }
          | update cells -c [real, user, kernel] { $"($in)sec" | into duration }
          | insert rows $rows
          | roll right
          | insert bin { if ($shellbin == $new_shellbin) { "new" } else { "old" } }
          | roll right
        }
      | flatten
  }
  | flatten
  | to nuon
```

This yields the following results

|bin|rows|rss_max|real|user|kernel|
|-|-|-|-|-|-|
|old|1000|44.1 MiB|110ms|90ms|30ms|
|old|10000|76.3 MiB|680ms|610ms|80ms|
|old|100000|396.5 MiB|5sec 460ms|5sec 80ms|680ms|
|new|1000|40.1 MiB|130ms|100ms|40ms|
|new|10000|45.2 MiB|590ms|550ms|90ms|
|new|100000|45.5 MiB|4sec 950ms|4sec 850ms|970ms|