@jntrnr 2023-10-11
Thanks for poking around and finding a way to get this fixed. I've been thinking about this the last couple days, and I wonder if we could fix this as part of an effort to adding streaming to Value. This is something we've talked about for at least the last year.

If Value had a `Value::Stream` form, then `from csv` could output it. That way, we're not materialising the whole output. Like you noticed, we also want to parse a stream in. This makes me wonder if what `open` should be doing is outputting a byte stream, and then `from` things could use it.

For the other part of the PR: I don't think we should have adding a new input type as part of this PR, specifically taking in rows. That feels out of scope and something that should be a separate PR.