@dead10ck 2023-10-12
> Thanks for poking around and finding a way to get this fixed. I've been thinking about this the last couple days, and I wonder if we could fix this as part of an effort to adding streaming to Value. This is something we've talked about for at least the last year.

Just to clarify, are you suggesting holding off on this implementation of the fix in favor of a more broad stroke to address streaming, or would you be interested in merging this first and then addressing the streaming more broadly?

> If Value had a `Value::Stream` form, then `from csv` could output it. That way, we're not materialising the whole output.

Sorry if I'm missing context, but isn't this what `PipelineData::ListStream` is? Adding a `Value::Stream` variant feels like the wrong place to me; to me, a "stream" is not really a data type, just a method of consumption. Like consider `Value::List`: I think it's a safe bet that most agree that this is a data type in its own right, even though it is composed of other `Value`s—it makes sense if you have a list of lists, and it makes sense if you have a record whose value is a list. But is the same true of a stream? What is a stream of streams? What does it mean if a record's value is a stream?

>Like you noticed, we also want to parse a stream in. This makes me wonder if what `open` should be doing is outputting a byte stream, and then `from` things could use it.

I think the fundamental conflict here is that nu wants to work with structured data, while everything outside of nu wants to work with bytes. Nu is powerful because it works with structured data natively, and yet it is absolutely essential to be able to interoperate with external code that only speaks raw byte streams for nu to be useful for very much of anything.

If `open` simply streams raw bytes, then there's not much difference between it and `cat`. But on the other hand, if structured data is our first class unit of data, then we need a way to translate the structured data to external code.

This is where I was thinking `std::io::Read` could come into play. Anything that can speak reading bytes through `Read` can interpret a `Value` directly. String and binary are the most trivially unambiguous native types, but there could be more than one reader that implements `Read` to provide an interface for the other types as well.

> For the other part of the PR: I don't think we should have adding a new input type as part of this PR, specifically taking in rows. That feels out of scope and something that should be a separate PR.

Sure thing, I've removed those parts. I'll open a separate PR.